unbiased wavelet power
Liu et al. 2007 suggest that the wavelet power in Terrence and Compo
needs to be divided by the scale to remove bias in the power spectrum.

Here this is implemented, as well as a calculation of the global wavelet
spectrum and a comparison between the biased and unbiased spectra as a
test case.